move the rest of the session apis into the apiSessions file
standardize and document apis
