+++ viewer/viewer.js
@@ -26,7 +26,6 @@ try {
   var Config = require('./config.js');
   var express = require('express');
   var stylus = require('stylus');
-  var util = require('util');
   var fs = require('fs');
   var fse = require('fs-ext');
   var async = require('async');
@@ -40,7 +39,6 @@ try {
   var version = require('./version');
   var http = require('http');
   var https = require('https');
-  var PNG = require('pngjs').PNG;
   var decode = require('./decode.js');
   var onHeaders = require('on-headers');
   var glob = require('glob');
@@ -49,7 +47,6 @@ try {
   var uuid = require('uuidv4').default;
   var RE2 = require('re2');
   var path = require('path');
-  var contentDisposition = require('content-disposition');
 } catch (e) {
   console.log("ERROR - Couldn't load some dependancies, maybe need to 'npm update' inside viewer directory", e);
   process.exit(1);
@@ -96,9 +93,9 @@ var compression = require('compression');
 // internal app deps
 let { internals } = require('./internals')(app, Config);
 let ViewerUtils = require('./viewerUtils')(Config, Db, molochparser, internals);
-let sessionAPIs = require('./apiSessions')(Config, Db, decode, internals, molochparser, Pcap, ViewerUtils);
+let sessionAPIs = require('./apiSessions')(Config, Db, decode, internals, molochparser, Pcap, version, ViewerUtils);
 let connectionAPIs = require('./apiConnections')(Config, Db, ViewerUtils, sessionAPIs);
-let statsAPIs = require('./apiStats')(Config, Db, internals, ViewerUtils); // TODO ECR
+let statsAPIs = require('./apiStats')(Config, Db, internals, ViewerUtils);
 
 // registers a get and a post
 app.getpost = (route, mw, func) => { app.get(route, mw, func); app.post(route, mw, func); };
@@ -229,8 +226,8 @@ app.use((req, res, next) => {
     return undefined;
   }` };
   mrc.reverseDNS = { category: 'ip', name: 'Get Reverse DNS', url: 'reverseDNS.txt?ip=%TEXT%', actionType: 'fetch' };
-  mrc.bodyHashMd5 = { category: 'md5', url: '%NODE%/%ID%/bodyHash/%TEXT%', name: 'Download File' };
-  mrc.bodyHashSha256 = { category: 'sha256', url: '%NODE%/%ID%/bodyHash/%TEXT%', name: 'Download File' };
+  mrc.bodyHashMd5 = { category: 'md5', url: 'api/sessions/%NODE%/%ID%/bodyHash/%TEXT%', name: 'Download File' };
+  mrc.bodyHashSha256 = { category: 'sha256', url: 'api/sessions/%NODE%/%ID%/bodyHash/%TEXT%', name: 'Download File' };
 
   for (var key in internals.rightClicks) {
     var rc = internals.rightClicks[key];
@@ -278,7 +275,7 @@ if (Config.get('passwordSecret')) {
       }
 
       // Don't look up user for receiveSession
-      if (req.url.match(/^\/receiveSession/)) {
+      if (req.url.match(/^\/receiveSession/) || req.url.match(/^\/api\/sessions\/receive/)) {
         return next();
       }
 
@@ -293,8 +290,8 @@ if (Config.get('passwordSecret')) {
       return;
     }
 
-    if (req.url.match(/^\/receiveSession/)) {
-      return res.send('receiveSession only allowed s2s');
+    if (req.url.match(/^\/receiveSession/) || req.url.match(/^\/api\/sessions\/receive/)) {
+      return res.send('receive session only allowed s2s');
     }
 
     function ucb (err, suser, userName) {
@@ -521,42 +518,6 @@ function createRightClicks () {
     });
   }, function () {
     internals.rightClicks = mrc;
-  });
-}
-
-// ============================================================================
-// REQUESTS
-// ============================================================================
-function makeRequest (node, path, user, cb) {
-  sessionAPIs.getViewUrl(node, function (err, viewUrl, client) {
-    let info = url.parse(viewUrl);
-    info.path = encodeURI(`${Config.basePath(node)}${path}`);
-    info.agent = (client === http ? internals.httpAgent : internals.httpsAgent);
-    info.timeout = 20 * 60 * 1000;
-    ViewerUtils.addAuth(info, user, node);
-    ViewerUtils.addCaTrust(info, node);
-
-    function responseFunc (pres) {
-      let response = '';
-      pres.on('data', function (chunk) {
-        response += chunk;
-      });
-      pres.on('end', function () {
-        cb(null, response);
-      });
-    }
-    let preq = client.request(info, responseFunc);
-    preq.on('error', (err) => {
-      // Try a second time on errors
-      console.log(`Retry ${info.path} on remote viewer: ${err}`);
-      let preq2 = client.request(info, responseFunc);
-      preq2.on('error', (err) => {
-        console.log(`Error with ${info.path} on remote viewer: ${err}`);
-        cb(err);
-      });
-      preq2.end();
-    });
-    preq.end();
   });
 }
 
@@ -1103,328 +1064,6 @@ function unshareView (req, res, user, sharedUser, endpoint, successMessage, erro
 }
 
 // session helpers ------------------------------------------------------------
-function processSessionIdAndDecode (id, numPackets, doneCb) {
-  var packets = [];
-  sessionAPIs.processSessionId(id, true, null, function (pcap, buffer, cb, i) {
-    var obj = {};
-    if (buffer.length > 16) {
-      pcap.decode(buffer, obj);
-    } else {
-      obj = { ip: { p: '' } };
-    }
-    packets[i] = obj;
-    cb(null);
-  },
-  function (err, session) {
-    if (err) {
-      console.log('ERROR - processSessionIdAndDecode', err);
-      return doneCb(err);
-    }
-    packets = packets.filter(Boolean);
-    if (packets.length === 0) {
-      return doneCb(null, session, []);
-    } else if (packets[0].ip === undefined) {
-      return doneCb(null, session, []);
-    } else if (packets[0].ip.p === 1) {
-      Pcap.reassemble_icmp(packets, numPackets, function (err, results) {
-        return doneCb(err, session, results);
-      });
-    } else if (packets[0].ip.p === 6) {
-      var key = session.srcIp;
-      Pcap.reassemble_tcp(packets, numPackets, key + ':' + session.srcPort, function (err, results) {
-        return doneCb(err, session, results);
-      });
-    } else if (packets[0].ip.p === 17) {
-      Pcap.reassemble_udp(packets, numPackets, function (err, results) {
-        return doneCb(err, session, results);
-      });
-    } else if (packets[0].ip.p === 132) {
-      Pcap.reassemble_sctp(packets, numPackets, function (err, results) {
-        return doneCb(err, session, results);
-      });
-    } else {
-      return doneCb(null, session, []);
-    }
-  },
-  numPackets, 10);
-}
-
-function localGetItemByHash (nodeName, sessionID, hash, cb) {
-  processSessionIdAndDecode(sessionID, 10000, function (err, session, incoming) {
-    if (err) {
-      return cb(err);
-    }
-    if (incoming.length === 0) {
-      return cb(null, null);
-    }
-    var options = {
-      id: sessionID,
-      nodeName: nodeName,
-      order: [],
-      'ITEM-HTTP': {
-        order: []
-      },
-      'ITEM-SMTP': {
-        order: ['BODY-UNBASE64']
-      },
-      'ITEM-HASH': {
-        hash: hash
-      },
-      'ITEM-CB': {
-      }
-    };
-
-    options.order.push('ITEM-HTTP');
-    options.order.push('ITEM-SMTP');
-    options.order.push('ITEM-HASH');
-    options.order.push('ITEM-CB');
-    options['ITEM-CB'].cb = function (err, items) {
-      if (err) {
-        return cb(err, null);
-      }
-      if (items === undefined || items.length === 0) {
-        return cb('No match', null);
-      }
-      return cb(err, items[0]);
-    };
-    decode.createPipeline(options, options.order, new decode.Pcap2ItemStream(options, incoming));
-  });
-}
-
-function writePcap (res, id, options, doneCb) {
-  var b = Buffer.alloc(0xfffe);
-  var nextPacket = 0;
-  var boffset = 0;
-  var packets = {};
-
-  sessionAPIs.processSessionId(id, false, function (pcap, buffer) {
-    if (options.writeHeader) {
-      res.write(buffer);
-      options.writeHeader = false;
-    }
-  },
-  function (pcap, buffer, cb, i) {
-    // Save this packet in its spot
-    packets[i] = buffer;
-
-    // Send any packets we have in order
-    while (packets[nextPacket]) {
-      buffer = packets[nextPacket];
-      delete packets[nextPacket];
-      nextPacket++;
-
-      if (boffset + buffer.length > b.length) {
-        res.write(b.slice(0, boffset));
-        boffset = 0;
-        b = Buffer.alloc(0xfffe);
-      }
-      buffer.copy(b, boffset, 0, buffer.length);
-      boffset += buffer.length;
-    }
-    cb(null);
-  },
-  function (err, session) {
-    if (err) {
-      console.trace('writePcap', err);
-      return doneCb(err);
-    }
-    res.write(b.slice(0, boffset));
-    doneCb(err);
-  }, undefined, 10);
-}
-
-function writePcapNg (res, id, options, doneCb) {
-  var b = Buffer.alloc(0xfffe);
-  var boffset = 0;
-
-  sessionAPIs.processSessionId(id, true, function (pcap, buffer) {
-    if (options.writeHeader) {
-      res.write(pcap.getHeaderNg());
-      options.writeHeader = false;
-    }
-  },
-  function (pcap, buffer, cb) {
-    if (boffset + buffer.length + 20 > b.length) {
-      res.write(b.slice(0, boffset));
-      boffset = 0;
-      b = Buffer.alloc(0xfffe);
-    }
-
-    /* Need to write the ng block, and conver the old timestamp */
-
-    b.writeUInt32LE(0x00000006, boffset); // Block Type
-    var len = ((buffer.length + 20 + 3) >> 2) << 2;
-    b.writeUInt32LE(len, boffset + 4); // Block Len 1
-    b.writeUInt32LE(0, boffset + 8); // Interface Id
-
-    // js has 53 bit numbers, this will over flow on Jun 05 2255
-    var time = buffer.readUInt32LE(0) * 1000000 + buffer.readUInt32LE(4);
-    b.writeUInt32LE(Math.floor(time / 0x100000000), boffset + 12); // Block Len 1
-    b.writeUInt32LE(time % 0x100000000, boffset + 16); // Interface Id
-
-    buffer.copy(b, boffset + 20, 8, buffer.length - 8); // cap_len, packet_len
-    b.fill(0, boffset + 12 + buffer.length, boffset + 12 + buffer.length + (4 - (buffer.length % 4)) % 4); // padding
-    boffset += len - 8;
-
-    b.writeUInt32LE(0, boffset); // Options
-    b.writeUInt32LE(len, boffset + 4); // Block Len 2
-    boffset += 8;
-
-    cb(null);
-  },
-  function (err, session) {
-    if (err) {
-      console.log('writePcapNg', err);
-      return;
-    }
-    res.write(b.slice(0, boffset));
-
-    session.version = version.version;
-    delete session.packetPos;
-    var json = JSON.stringify(session);
-
-    var len = ((json.length + 20 + 3) >> 2) << 2;
-    b = Buffer.alloc(len);
-
-    b.writeUInt32LE(0x80808080, 0); // Block Type
-    b.writeUInt32LE(len, 4); // Block Len 1
-    b.write('MOWL', 8); // Magic
-    b.writeUInt32LE(json.length, 12); // Block Len 1
-    b.write(json, 16); // Magic
-    b.fill(0, 16 + json.length, 16 + json.length + (4 - (json.length % 4)) % 4); // padding
-    b.writeUInt32LE(len, len - 4); // Block Len 2
-    res.write(b);
-
-    doneCb(err);
-  });
-}
-
-function sessionsPcapList (req, res, list, pcapWriter, extension) {
-  if (list.length > 0 && list[0].fields) {
-    list = list.sort(function (a, b) { return a.fields.lastPacket - b.fields.lastPacket; });
-  } else if (list.length > 0 && list[0]._source) {
-    list = list.sort(function (a, b) { return a._source.lastPacket - b._source.lastPacket; });
-  }
-
-  var options = { writeHeader: true };
-
-  async.eachLimit(list, 10, function (item, nextCb) {
-    var fields = item._source || item.fields;
-    sessionAPIs.isLocalView(fields.node, function () {
-      // Get from our DISK
-      pcapWriter(res, Db.session2Sid(item), options, nextCb);
-    },
-    function () {
-      // Get from remote DISK
-      sessionAPIs.getViewUrl(fields.node, function (err, viewUrl, client) {
-        var buffer = Buffer.alloc(fields.totPackets * 20 + fields.totBytes);
-        var bufpos = 0;
-        var info = url.parse(viewUrl);
-        info.path = Config.basePath(fields.node) + fields.node + '/' + extension + '/' + Db.session2Sid(item) + '.' + extension;
-        info.agent = (client === http ? internals.httpAgent : internals.httpsAgent);
-
-        ViewerUtils.addAuth(info, req.user, fields.node);
-        ViewerUtils.addCaTrust(info, fields.node);
-        var preq = client.request(info, function (pres) {
-          pres.on('data', function (chunk) {
-            if (bufpos + chunk.length > buffer.length) {
-              var tmp = Buffer.alloc(buffer.length + chunk.length * 10);
-              buffer.copy(tmp, 0, 0, bufpos);
-              buffer = tmp;
-            }
-            chunk.copy(buffer, bufpos);
-            bufpos += chunk.length;
-          });
-          pres.on('end', function () {
-            if (bufpos < 24) {
-            } else if (options.writeHeader) {
-              options.writeHeader = false;
-              res.write(buffer.slice(0, bufpos));
-            } else {
-              res.write(buffer.slice(24, bufpos));
-            }
-            setImmediate(nextCb);
-          });
-        });
-        preq.on('error', function (e) {
-          console.log("ERROR - Couldn't proxy pcap request=", info, '\nerror=', e);
-          nextCb(null);
-        });
-        preq.end();
-      });
-    });
-  }, function (err) {
-    res.end();
-  });
-}
-
-function sessionsPcap (req, res, pcapWriter, extension) {
-  ViewerUtils.noCache(req, res, 'application/vnd.tcpdump.pcap');
-
-  if (req.query.ids) {
-    var ids = ViewerUtils.queryValueToArray(req.query.ids);
-
-    sessionAPIs.sessionsListFromIds(req, ids, ['lastPacket', 'node', 'totBytes', 'totPackets', 'rootId'], function (err, list) {
-      sessionsPcapList(req, res, list, pcapWriter, extension);
-    });
-  } else {
-    sessionAPIs.sessionsListFromQuery(req, res, ['lastPacket', 'node', 'totBytes', 'totPackets', 'rootId'], function (err, list) {
-      sessionsPcapList(req, res, list, pcapWriter, extension);
-    });
-  }
-}
-
-function reqGetRawBody (req, cb) {
-  processSessionIdAndDecode(req.params.id, 10000, function (err, session, incoming) {
-    if (err) {
-      return cb(err);
-    }
-
-    if (incoming.length === 0) {
-      return cb(null, null);
-    }
-
-    var options = {
-      id: session.id,
-      nodeName: req.params.nodeName,
-      order: [],
-      'ITEM-HTTP': {
-        order: []
-      },
-      'ITEM-SMTP': {
-        order: ['BODY-UNBASE64']
-      },
-      'ITEM-CB': {
-      },
-      'ITEM-RAWBODY': {
-        bodyNumber: +req.params.bodyNum
-      }
-    };
-
-    if (req.query.needgzip) {
-      options['ITEM-HTTP'].order.push('BODY-UNCOMPRESS');
-      options['ITEM-SMTP'].order.push('BODY-UNCOMPRESS');
-    }
-
-    options.order.push('ITEM-HTTP');
-    options.order.push('ITEM-SMTP');
-
-    options.order.push('ITEM-RAWBODY');
-    options.order.push('ITEM-CB');
-    options['ITEM-CB'].cb = function (err, items) {
-      if (err) {
-        return cb(err);
-      }
-      if (items === undefined || items.length === 0) {
-        return cb('No match');
-      }
-      cb(err, items[0].data);
-    };
-
-    decode.createPipeline(options, options.order, new decode.Pcap2ItemStream(options, incoming));
-  });
-}
-
 function sendSessionWorker (options, cb) {
   var packetslen = 0;
   var packets = [];
@@ -1491,7 +1130,7 @@ function sendSessionWorker (options, cb) {
       return cb();
     }
 
-    var info = url.parse(sobj.url + '/receiveSession?saveId=' + options.saveId);
+    var info = url.parse(sobj.url + '/api/sessions/receive?saveId=' + options.saveId);
     ViewerUtils.addAuth(info, options.user, options.nodeName, sobj.serverSecret || sobj.passwordSecret);
     info.method = 'POST';
 
@@ -1529,41 +1168,6 @@ function sendSessionWorker (options, cb) {
 }
 
 internals.sendSessionQueue = async.queue(sendSessionWorker, 10);
-
-function sendSessionsList (req, res, list) {
-  if (!list) { return res.molochError(200, 'Missing list of sessions'); }
-
-  var saveId = Config.nodeName() + '-' + new Date().getTime().toString(36);
-
-  async.eachLimit(list, 10, function (item, nextCb) {
-    var fields = item._source || item.fields;
-    let sid = Db.session2Sid(item);
-    sessionAPIs.isLocalView(fields.node, function () {
-      var options = {
-        user: req.user,
-        cluster: req.body.cluster,
-        id: sid,
-        saveId: saveId,
-        tags: req.body.tags,
-        nodeName: fields.node
-      };
-      // Get from our DISK
-      internals.sendSessionQueue.push(options, nextCb);
-    },
-    function () {
-      let path = `${fields.node}/sendSession/${sid}?saveId=${saveId}&cluster=${req.body.cluster}`;
-      if (req.body.tags) {
-        path += `&tags=${req.body.tags}`;
-      }
-
-      makeRequest(fields.node, path, req.user, (err, response) => {
-        setImmediate(nextCb);
-      });
-    });
-  }, function (err) {
-    return res.end(JSON.stringify({ success: true, text: 'Sending of ' + list.length + ' sessions complete' }));
-  });
-}
 
 var qlworking = {};
 function sendSessionsListQL (pOptions, list, nextQLCb) {
@@ -1603,10 +1207,10 @@ function sendSessionsListQL (pOptions, list, nextQLCb) {
     },
     function () {
       // Get from remote DISK
-      sessionAPIs.getViewUrl(node, function (err, viewUrl, client) {
+      ViewerUtils.getViewUrl(node, function (err, viewUrl, client) {
         var info = url.parse(viewUrl);
         info.method = 'POST';
-        info.path = Config.basePath(node) + node + '/sendSessions?saveId=' + pOptions.saveId + '&cluster=' + pOptions.cluster;
+        info.path = `api/sessions/${Config.basePath(node) + node}/send?saveId=${pOptions.saveId}&cluster=${pOptions.cluster}`;
         info.agent = (client === http ? internals.httpAgent : internals.httpsAgent);
         if (pOptions.tags) {
           info.path += '&tags=' + pOptions.tags;
@@ -1759,7 +1363,7 @@ function packetSearch (packet, options) {
 
 function sessionHunt (sessionId, options, cb) {
   if (options.type === 'reassembled') {
-    processSessionIdAndDecode(sessionId, options.size || 10000, function (err, session, packets) {
+    sessionAPIs.processSessionIdAndDecode(sessionId, options.size || 10000, function (err, session, packets) {
       if (err) {
         return cb(null, false);
       }
@@ -1972,7 +1576,7 @@ function huntFailedSessions (hunt, huntId, options, searchedSessions, user) {
 
       session = session._source;
 
-      makeRequest(session.node, `${session.node}/hunt/${huntId}/remote/${sessionId}`, user, (err, response) => {
+      ViewerUtils.makeRequest(session.node, `${session.node}/hunt/${huntId}/remote/${sessionId}`, user, (err, response) => {
         if (err) {
           return continueHuntSkipSession(hunt, huntId, session, sessionId, searchedSessions, cb);
         }
@@ -2085,7 +1689,7 @@ function runHuntJob (huntId, hunt, query, user) {
       function () { // Check Remotely
         let path = `${node}/hunt/${huntId}/remote/${sessionId}`;
 
-        makeRequest(node, path, user, (err, response) => {
+        ViewerUtils.makeRequest(node, path, user, (err, response) => {
           if (err) {
             return continueHuntSkipSession(hunt, huntId, session, sessionId, searchedSessions, cb);
           }
@@ -2460,7 +2064,7 @@ function scrubList (req, res, whatToRemove, list) {
     function () {
       // Get from remote DISK
       let path = `${fields.node}/delete/${whatToRemove}/${Db.session2Sid(item)}`;
-      makeRequest(fields.node, path, req.user, function (err, response) {
+      ViewerUtils.makeRequest(fields.node, path, req.user, function (err, response) {
         setImmediate(nextCb);
       });
     });
@@ -4157,173 +3761,6 @@ app.get('/:nodeName/:fileNum/filesize.json', [noCacheJson, checkPermissions(['hi
   });
 });
 
-// Get a file given a hash of that file
-app.get('/bodyHash/:hash', logAction('bodyhash'), function (req, res) {
-  var hash = null;
-  var nodeName = null;
-  var sessionID = null;
-
-  sessionAPIs.buildSessionQuery(req, (bsqErr, query, indices) => {
-    if (bsqErr) {
-      res.status(400);
-      return res.end(bsqErr);
-    }
-
-    query.size = 1;
-    query.sort = { lastPacket: { order: 'desc' } };
-    query._source = ['node'];
-
-    if (Config.debug) {
-      console.log(`sessions.json ${indices} query`, JSON.stringify(query, null, 1));
-    }
-    Db.searchPrimary(indices, 'session', query, null, function (err, sessions) {
-      if (err) {
-        console.log('Error -> Db Search ', err);
-        res.status(400);
-        res.end(err);
-      } else if (sessions.error) {
-        console.log('Error -> Db Search ', sessions.error);
-        res.status(400);
-        res.end(sessions.error);
-      } else {
-        if (Config.debug) {
-          console.log('bodyHash result', util.inspect(sessions, false, 50));
-        }
-        if (sessions.hits.hits.length > 0) {
-          nodeName = sessions.hits.hits[0]._source.node;
-          sessionID = Db.session2Sid(sessions.hits.hits[0]);
-          hash = req.params.hash;
-
-          sessionAPIs.isLocalView(nodeName, function () { // get file from the local disk
-            localGetItemByHash(nodeName, sessionID, hash, (err, item) => {
-              if (err) {
-                res.status(400);
-                return res.end(err);
-              } else if (item) {
-                ViewerUtils.noCache(req, res, 'application/force-download');
-                res.setHeader('content-disposition', contentDisposition(item.bodyName + '.pellet'));
-                return res.end(item.data);
-              } else {
-                res.status(400);
-                return res.end('No Match');
-              }
-            });
-          },
-          function () { // get file from the remote disk
-            let preq = Object.assign({}, req);
-            preq.params.nodeName = nodeName;
-            preq.params.id = sessionID;
-            preq.params.hash = hash;
-            preq.url = Config.basePath(nodeName) + nodeName + '/' + sessionID + '/bodyHash/' + hash;
-            return sessionAPIs.proxyRequest(preq, res);
-          });
-        } else {
-          res.status(400);
-          res.end('No Match Found');
-        }
-      }
-    });
-  });
-});
-
-app.get('/:nodeName/:id/bodyHash/:hash', checkProxyRequest, function (req, res) {
-  localGetItemByHash(req.params.nodeName, req.params.id, req.params.hash, (err, item) => {
-    if (err) {
-      res.status(400);
-      return res.end(err);
-    } else if (item) {
-      ViewerUtils.noCache(req, res, 'application/force-download');
-      res.setHeader('content-disposition', contentDisposition(item.bodyName + '.pellet'));
-      return res.end(item.data);
-    } else {
-      res.status(400);
-      return res.end('No Match');
-    }
-  });
-});
-
-app.get('/:nodeName/pcapng/:id.pcapng', [checkProxyRequest, checkPermissions(['disablePcapDownload'])], (req, res) => {
-  ViewerUtils.noCache(req, res, 'application/vnd.tcpdump.pcap');
-  writePcapNg(res, req.params.id, { writeHeader: !req.query || !req.query.noHeader || req.query.noHeader !== 'true' }, function () {
-    res.end();
-  });
-});
-
-app.get('/:nodeName/pcap/:id.pcap', [checkProxyRequest, checkPermissions(['disablePcapDownload'])], (req, res) => {
-  ViewerUtils.noCache(req, res, 'application/vnd.tcpdump.pcap');
-
-  writePcap(res, req.params.id, { writeHeader: !req.query || !req.query.noHeader || req.query.noHeader !== 'true' }, function () {
-    res.end();
-  });
-});
-
-app.get('/:nodeName/raw/:id.png', [checkProxyRequest, checkPermissions(['disablePcapDownload'])], function (req, res) {
-  ViewerUtils.noCache(req, res, 'image/png');
-
-  processSessionIdAndDecode(req.params.id, 1000, function (err, session, results) {
-    if (err) {
-      return res.send(internals.emptyPNG);
-    }
-    var size = 0;
-    var i, ilen;
-    for (i = (req.query.type !== 'dst' ? 0 : 1), ilen = results.length; i < ilen; i += 2) {
-      size += results[i].data.length + 2 * internals.PNG_LINE_WIDTH - (results[i].data.length % internals.PNG_LINE_WIDTH);
-    }
-    var buffer = Buffer.alloc(size, 0);
-    var pos = 0;
-    if (size === 0) {
-      return res.send(internals.emptyPNG);
-    }
-    for (i = (req.query.type !== 'dst' ? 0 : 1), ilen = results.length; i < ilen; i += 2) {
-      results[i].data.copy(buffer, pos);
-      pos += results[i].data.length;
-      var fillpos = pos;
-      pos += 2 * internals.PNG_LINE_WIDTH - (results[i].data.length % internals.PNG_LINE_WIDTH);
-      buffer.fill(0xff, fillpos, pos);
-    }
-
-    var png = new PNG({ width: internals.PNG_LINE_WIDTH, height: (size / internals.PNG_LINE_WIDTH) - 1 });
-    png.data = buffer;
-    res.send(PNG.sync.write(png, { inputColorType: 0, colorType: 0, bitDepth: 8, inputHasAlpha: false }));
-  });
-});
-
-app.get('/:nodeName/raw/:id', [checkProxyRequest, checkPermissions(['disablePcapDownload'])], function (req, res) {
-  ViewerUtils.noCache(req, res, 'application/vnd.tcpdump.pcap');
-
-  processSessionIdAndDecode(req.params.id, 10000, function (err, session, results) {
-    if (err) {
-      return res.send('Error');
-    }
-    for (let i = (req.query.type !== 'dst' ? 0 : 1), ilen = results.length; i < ilen; i += 2) {
-      res.write(results[i].data);
-    }
-    res.end();
-  });
-});
-
-app.get('/:nodeName/entirePcap/:id.pcap', [checkProxyRequest, checkPermissions(['disablePcapDownload'])], (req, res) => {
-  ViewerUtils.noCache(req, res, 'application/vnd.tcpdump.pcap');
-
-  var options = { writeHeader: true };
-
-  var query = { _source: ['rootId'],
-    size: 1000,
-    query: { term: { rootId: req.params.id } },
-    sort: { lastPacket: { order: 'asc' } }
-  };
-
-  console.log('entirePcap query', JSON.stringify(query));
-
-  Db.searchPrimary('sessions2-*', 'session', query, null, function (err, data) {
-    async.forEachSeries(data.hits.hits, function (item, nextCb) {
-      writePcap(res, Db.session2Sid(item), options, nextCb);
-    }, function (err) {
-      res.end();
-    });
-  });
-});
-
 // misc apis ------------------------------------------------------------------
 app.get('/titleconfig', checkPermissions(['webEnabled']), (req, res) => {
   var titleConfig = Config.get('titleTemplate', '_cluster_ - _page_ _-view_ _-expression_');
@@ -4381,6 +3818,15 @@ app.get('/molochRightClick', [noCacheJson, checkPermissions(['webEnabled'])], (r
 app.get(['/api/eshealth', '/eshealth.json'], [noCacheJson], (req, res) => {
   Db.healthCache(function (err, health) {
     res.send(health);
+  });
+});
+
+app.get('/reverseDNS.txt', [noCacheJson, logAction()], (req, res) => {
+  dns.reverse(req.query.ip, (err, data) => {
+    if (err) {
+      return res.send('reverse error');
+    }
+    return res.send(data.join(', '));
   });
 });
 
@@ -4662,266 +4108,100 @@ app.post( // remove tags endpoint
   sessionAPIs.removeTags
 );
 
-app.get('/:nodeName/:id/body/:bodyType/:bodyNum/:bodyName', checkProxyRequest, function (req, res) {
-  reqGetRawBody(req, function (err, data) {
-    if (err) {
-      console.trace(err);
-      return res.end('Error');
-    }
-    res.setHeader('Content-Type', 'application/force-download');
-    res.setHeader('content-disposition', contentDisposition(req.params.bodyName));
-    return res.end(data);
-  });
-});
-
-app.get('/:nodeName/:id/bodypng/:bodyType/:bodyNum/:bodyName', checkProxyRequest, function (req, res) {
-  reqGetRawBody(req, function (err, data) {
-    if (err || data === null || data.length === 0) {
-      return res.send(internals.emptyPNG);
-    }
-    res.setHeader('Content-Type', 'image/png');
-
-    var png = new PNG({ width: internals.PNG_LINE_WIDTH, height: Math.ceil(data.length / internals.PNG_LINE_WIDTH) });
-    png.data = data;
-    res.send(PNG.sync.write(png, { inputColorType: 0, colorType: 0, bitDepth: 8, inputHasAlpha: false }));
-  });
-});
-
-app.get(/\/sessions.pcapng.*/, [logAction(), checkPermissions(['disablePcapDownload'])], (req, res) => {
-  return sessionsPcap(req, res, writePcapNg, 'pcapng');
-});
-
-/**
- * GET - /api/sessions/pcap
- *
- * Retrieve the raw session data in pcap format
- * @name sessions/pcap
- * @param {SessionsQuery} query - The request query to filter sessions
- * @param {string} ids - The list of ids to return
- * @param {boolean} segments=false - When set return linked segments
- * @returns {pcap} A PCAP file with the sessions requested
- */
-app.get(
-  ['/api/sessions/pcap', /\/sessions.pcap.*/],
-  [logAction(), checkPermissions(['disablePcapDownload'])],
-  (req, res) => {
-    return sessionsPcap(req, res, writePcap, 'pcap');
-  }
+app.get( // session body file endpoint
+  ['/api/sessions/:nodeName/:id/body/:bodyType/:bodyNum/:bodyName', '/:nodeName/:id/body/:bodyType/:bodyNum/:bodyName'],
+  [checkProxyRequest],
+  sessionAPIs.getRawBody
 );
 
-app.get('/:nodeName/sendSession/:id', checkProxyRequest, function (req, res) {
-  ViewerUtils.noCache(req, res);
-  res.statusCode = 200;
+app.get( // session body file image endpoint
+  ['/api/sessions/:nodeName/:id/bodypng/:bodyType/:bodyNum/:bodyName', '/:nodeName/:id/bodypng/:bodyType/:bodyNum/:bodyName'],
+  [checkProxyRequest],
+  sessionAPIs.getFilePNG
+);
 
-  var options = {
-    user: req.user,
-    cluster: req.query.cluster,
-    id: req.params.id,
-    saveId: req.query.saveId,
-    tags: req.body.tags,
-    nodeName: req.params.nodeName
-  };
+app.get( // session pcap endpoint
+  ['/api/sessions/pcap', /\/sessions.pcap.*/],
+  [logAction(), checkPermissions(['disablePcapDownload'])],
+  sessionAPIs.getPCAP
+);
 
-  internals.sendSessionQueue.push(options, function () {
-    res.end();
-  });
-});
+app.get( // session pcapng endpoint
+  ['/api/sessions/pcapng', /\/sessions.pcapng.*/],
+  [logAction(), checkPermissions(['disablePcapDownload'])],
+  sessionAPIs.getPCAPNG
+);
 
-app.post('/:nodeName/sendSessions', checkProxyRequest, function (req, res) {
-  ViewerUtils.noCache(req, res);
-  res.statusCode = 200;
+app.get( // session node pcap endpoint
+  ['/api/sessions/:nodeName/pcap/:id.pcap', '/:nodeName/pcap/:id.pcap'],
+  [checkProxyRequest, checkPermissions(['disablePcapDownload'])],
+  sessionAPIs.getPCAPFromNode
+);
 
-  if (req.body.ids === undefined ||
-      req.query.cluster === undefined ||
-      req.query.saveId === undefined ||
-      req.body.tags === undefined) {
-    return res.end();
-  }
+app.get( // session node pcapng endpoint
+  ['/api/sessions/:nodeName/pcapng/:id.pcapng', '/:nodeName/pcapng/:id.pcapng'],
+  [checkProxyRequest, checkPermissions(['disablePcapDownload'])],
+  sessionAPIs.getPCAPNGFromNode
+);
 
-  var count = 0;
-  var ids = ViewerUtils.queryValueToArray(req.body.ids);
-  ids.forEach(function (id) {
-    var options = {
-      user: req.user,
-      cluster: req.query.cluster,
-      id: id,
-      saveId: req.query.saveId,
-      tags: req.body.tags,
-      nodeName: req.params.nodeName
-    };
+app.get( // session entire pcap endpoint
+  ['/api/sessions/:nodeName/entirepcap/:id.pcap', '/:nodeName/entirePcap/:id.pcap'],
+  [checkProxyRequest, checkPermissions(['disablePcapDownload'])],
+  sessionAPIs.getEntirePCAP
+);
 
-    count++;
-    internals.sendSessionQueue.push(options, function () {
-      count--;
-      if (count === 0) {
-        return res.end();
-      }
-    });
-  });
-});
+app.get( // session packets file image endpoint
+  ['/api/sessions/:nodeName/raw/:id.png', '/:nodeName/raw/:id.png'],
+  [checkProxyRequest, checkPermissions(['disablePcapDownload'])],
+  sessionAPIs.getPacketPNG
+);
 
-app.post('/receiveSession', [noCacheJson], function receiveSession (req, res) {
-  if (!req.query.saveId) { return res.molochError(200, 'Missing saveId'); }
+app.get( // session raw packets endpoint
+  ['/api/sessions/:nodeName/raw/:id', '/:nodeName/raw/:id'],
+  [checkProxyRequest, checkPermissions(['disablePcapDownload'])],
+  sessionAPIs.getRawPackets
+);
 
-  req.query.saveId = req.query.saveId.replace(/[^-a-zA-Z0-9_]/g, '');
+app.get( // session file bodyhash endpoint
+  ['/api/sessions/bodyhash/:hash', '/bodyHash/:hash'],
+  [logAction('bodyhash')],
+  sessionAPIs.getBodyHash
+);
 
-  // JS Static Variable :)
-  receiveSession.saveIds = receiveSession.saveIds || {};
+app.get( // session file bodyhash endpoint
+  ['/api/sessions/:nodeName/:id/bodyhash/:hash', '/:nodeName/:id/bodyHash/:hash'],
+  [checkProxyRequest],
+  sessionAPIs.getBodyHashFromNode
+);
 
-  var saveId = receiveSession.saveIds[req.query.saveId];
-  if (!saveId) {
-    saveId = receiveSession.saveIds[req.query.saveId] = { start: 0 };
-  }
+app.get( // sessions get decodings endpoint
+  ['/api/sessions/decodings', '/decodings'],
+  [noCacheJson],
+  sessionAPIs.getDecodings
+);
 
-  var sessionlen = -1;
-  var filelen = -1;
-  var written = 0;
-  var session = null;
-  var buffer;
-  var file;
-  var writeHeader;
+app.get( // session send to node endpoint
+  ['/api/sessions/:nodeName/send/:id', '/:nodeName/sendSession/:id'],
+  [checkProxyRequest],
+  sessionAPIs.sendSessionToNode
+);
 
-  function makeFilename (cb) {
-    if (saveId.filename) {
-      return cb(saveId.filename);
-    }
+app.post( // sessions send to node endpoint
+  ['/api/sessions/:nodeName/send', '/:nodeName/sendSessions'],
+  [checkProxyRequest],
+  sessionAPIs.sendSessionsToNode
+);
 
-    // Just keep calling ourselves every 100 ms until we have a filename
-    if (saveId.inProgress) {
-      return setTimeout(makeFilename, 100, cb);
-    }
+app.post( // sessions send endpoint
+  ['/api/sessions/send', '/sendSessions'],
+  sessionAPIs.sendSessions
+);
 
-    saveId.inProgress = 1;
-    Db.getSequenceNumber('fn-' + Config.nodeName(), function (err, seq) {
-      var filename = Config.get('pcapDir') + '/' + Config.nodeName() + '-' + seq + '-' + req.query.saveId + '.pcap';
-      saveId.seq = seq;
-      Db.indexNow('files', 'file', Config.nodeName() + '-' + saveId.seq, { num: saveId.seq, name: filename, first: session.firstPacket, node: Config.nodeName(), filesize: -1, locked: 1 }, function () {
-        cb(filename);
-        saveId.filename = filename; // Don't set the saveId.filename until after the first request completes its callback.
-      });
-    });
-  }
-
-  function saveSession () {
-    var id = session.id;
-    delete session.id;
-    Db.indexNow(Db.sid2Index(id), 'session', Db.sid2Id(id), session, function (err, info) {
-    });
-  }
-
-  function chunkWrite (chunk) {
-    // Write full chunk if first packet and writeHeader or not first packet
-    if (writeHeader || written !== 0) {
-      writeHeader = false;
-      file.write(chunk);
-    } else {
-      file.write(chunk.slice(24));
-    }
-    written += chunk.length; // Pretend we wrote it all
-  }
-
-  req.on('data', function (chunk) {
-    // If the file is open, just write the current chunk
-    if (file) {
-      return chunkWrite(chunk);
-    }
-
-    // If no file is open, then save the current chunk to the end of the buffer.
-    if (!buffer) {
-      buffer = chunk;
-    } else {
-      buffer = Buffer.concat([buffer, chunk]);
-    }
-
-    // Found the lengths
-    if (sessionlen === -1 && (buffer.length >= 12)) {
-      sessionlen = buffer.readUInt32BE(0);
-      filelen = buffer.readUInt32BE(8);
-      buffer = buffer.slice(12);
-    }
-
-    // If we know the session len and haven't read the session
-    if (sessionlen !== -1 && !session && buffer.length >= sessionlen) {
-      session = JSON.parse(buffer.toString('utf8', 0, sessionlen));
-      session.node = Config.nodeName();
-      buffer = buffer.slice(sessionlen);
-
-      if (filelen > 0) {
-        req.pause();
-
-        makeFilename(function (filename) {
-          req.resume();
-          session.packetPos[0] = -saveId.seq;
-          session.fileId = [saveId.seq];
-
-          if (saveId.start === 0) {
-            file = fs.createWriteStream(filename, { flags: 'w' });
-          } else {
-            file = fs.createWriteStream(filename, { start: saveId.start, flags: 'r+' });
-          }
-          writeHeader = saveId.start === 0;
-
-          // Adjust packet location based on where we start writing
-          if (saveId.start > 0) {
-            for (var p = 1, plen = session.packetPos.length; p < plen; p++) {
-              session.packetPos[p] += (saveId.start - 24);
-            }
-          }
-
-          // Filelen always includes header, if we don't write header subtract it
-          saveId.start += filelen;
-          if (!writeHeader) {
-            saveId.start -= 24;
-          }
-
-          // Still more data in buffer, start of pcap
-          if (buffer.length > 0) {
-            chunkWrite(buffer);
-          }
-
-          saveSession();
-        });
-      } else {
-        saveSession();
-      }
-    }
-  });
-
-  req.on('end', function (chunk) {
-    if (file) {
-      file.end();
-    }
-    return res.send({ success: true });
-  });
-});
-
-app.post('/sendSessions', function (req, res) {
-  if (req.body.ids) {
-    var ids = ViewerUtils.queryValueToArray(req.body.ids);
-
-    sessionAPIs.sessionsListFromIds(req, ids, ['node'], function (err, list) {
-      sendSessionsList(req, res, list);
-    });
-  } else {
-    sessionAPIs.sessionsListFromQuery(req, res, ['node'], function (err, list) {
-      sendSessionsList(req, res, list);
-    });
-  }
-});
-
-app.get('/decodings', [noCacheJson], function (req, res) {
-  var decodeItems = decode.settings();
-  res.send(JSON.stringify(decodeItems));
-});
-
-app.get('/reverseDNS.txt', [noCacheJson, logAction()], (req, res) => {
-  dns.reverse(req.query.ip, (err, data) => {
-    if (err) {
-      return res.send('reverse error');
-    }
-    return res.send(data.join(', '));
-  });
-});
+app.post( // sessions recieve endpoint
+  ['/api/sessions/receive', '/receiveSession'],
+  [noCacheJson],
+  sessionAPIs.receiveSession
+);
 
 // connections apis -----------------------------------------------------------
 app.getpost( // connections endpoint (POST or GET) - uses fillQueryFromBody to
@@ -5611,7 +4891,7 @@ app.post('/upload', [checkCookieToken, multer({ dest: '/tmp', limits: internals.
 // cyberchef apis -------------------------------------------------------------
 // loads the src or dst packets for a session and sends them to cyberchef
 app.get('/cyberchef/:nodeName/session/:id', checkPermissions(['webEnabled']), checkProxyRequest, unsafeInlineCspHeader, (req, res) => {
-  processSessionIdAndDecode(req.params.id, 10000, function (err, session, results) {
+  sessionAPIs.processSessionIdAndDecode(req.params.id, 10000, function (err, session, results) {
     if (err) {
       console.log(`ERROR - /${req.params.nodeName}/session/${req.params.id}/cyberchef`, err);
       return res.end('Error - ' + err);
